Rename `viewport_resolution` to `viewport_size` and move it to `[render]`
At a logical level, that's where it belongs, next to the `aspect`
setting it's closely related to. Managing the viewport size has nothing
to do with interfacing with the host OS, so it was a mistake to put it
into the `[sdl]` section in the first place.

Secondly, the name `viewport_resolution` was an unfortunate choice in
this age of high-DPI displays as `WxH` settings are interpreted as
logical units, and "resolution" implies physical resolution in pixels
(as in "display resolution" of a monitor).

Lastly, changing `viewport_resolution` triggered re-creating the window
and resetting the window size in windowed mode, and flashing and/or
delays in fullscreen. This is the current behaviour when changing *any*
setting in the `[sdl]` section. While that issue should be addressed
separately, moving `viewport_resolution` into `[render]` at least fixes
the problem for this single setting only.
